requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
transformers==4.36.2
torch==2.1.2
faster-whisper==1.0.3
sentence-transformers==2.2.2

